better zustand management
